Ultra sonic done and waiting for millis
The ultrasonic part of the robot is finished for now as this was properly tested, I use float for a better accuracy.
Right now I am waiting for Gripper with Millis as without those i cannot make delays using the millis in a way that it wouldn't crash/stop the bot.